Split redux actions in several files
