+++ 06 Redux/src/actions/doctors/assignDoctorsAction.ts
@@ -1,0 +1,12 @@
+import { Action, ActionCreator } from 'redux';
+
+export const ASSIGN_DOCTORS: string = "ASSIGN_DOCTORS";
+export interface AssignDoctorsAction extends Action {
+  doctors: Array<string>;
+}
+
+export const assignDoctors: ActionCreator<AssignDoctorsAction> =
+(doctors: Array<string>) => ({
+  type: ASSIGN_DOCTORS,
+  doctors
+});
