+++ 06 Redux/src/actions/doctors/loadDoctorsAction.ts
@@ -1,0 +1,10 @@
+import { patientAPI } from '../../api/patientAPI';
+import { assignDoctors } from './assignDoctorsAction';
+
+export const loadDoctors = () => {
+  return dispatcher => {
+    patientAPI.getAllDoctorsAsync().then((doctors: Array<string>) => {
+      dispatcher(assignDoctors(doctors));
+    });
+  }
+}
