+++ 06 Redux/src/actions/patient/assignPatientAction.ts
@@ -1,0 +1,13 @@
+import { Action, ActionCreator } from 'redux';
+import { Patient } from '../../model/patient';
+
+export const ASSIGN_PATIENT: string = "ASSIGN_PATIENT";
+export interface AssignPatientAction extends Action {
+  patient: Patient;
+}
+
+export const assignPatient: ActionCreator<AssignPatientAction> =
+(patient: Patient) => ({
+  type: ASSIGN_PATIENT,
+  patient
+});
