+++ 06 Redux/src/components/patients/patientList.container.ts
@@ -3,7 +3,7 @@ import { Patient } from '../../model/patient';
 import { Store } from 'redux';
 import { AppStore } from '../../store';
 import { AppState } from '../../reducers/';
-import { loadPatients } from '../../actions/patientsActions';
+import { loadPatients } from '../../actions/patients/loadPatientsAction';
 
 @Component({
   selector: 'patient-list-container',
