+++ 06 Redux/src/components/patients/searchPatient.container.ts
@@ -2,7 +2,7 @@ import { Component, Inject } from '@angular/core';
 import { Store } from 'redux';
 import { AppStore } from '../../store';
 import { AppState } from '../../reducers/';
-import { loadSpecialties } from '../../actions/specialtiesActions';
+import { loadSpecialties } from '../../actions/specialties/loadSpecialtiesAction';
 
 @Component({
   selector: 'search-patient-container',
