+++ 06 Redux/src/reducers/doctorsReducer.ts
@@ -1,5 +1,5 @@
 import { Reducer } from 'redux';
-import { ASSIGN_DOCTORS, AssignDoctorsAction } from '../actions/doctorsActions';
+import { ASSIGN_DOCTORS, AssignDoctorsAction } from '../actions/doctors/assignDoctorsAction';
 
 export const doctorsReducer: Reducer<Array<string>> =
 (state: Array<string> = [], action: AssignDoctorsAction): Array<string> => {
