+++ 06 Redux/src/reducers/patientsReducer.ts
@@ -1,6 +1,6 @@
 import { Reducer } from 'redux';
 import { Patient } from '../model/patient';
-import { ASSIGN_PATIENTS, AssignPatientsAction } from '../actions/patientsActions';
+import { ASSIGN_PATIENTS, AssignPatientsAction } from '../actions/patients/assignPatientsAction';
 
 export const patientsReducer: Reducer<Array<Patient>> =
 (state: Array<Patient> = [], action: AssignPatientsAction): Array<Patient> => {
